moving data_table and sql out of main explorer app script
