+++ neurobooth_explorer/__init__.py
@@ -1,0 +1,3 @@
+"""Neurobooth Explorer"""
+
+__version__ = "0.1.dev0"
